Huge amount of fixes

+++ JsonModels/RoundData.cs
@@ -5,6 +5,6 @@
         public int? RoundNumber { get; set; }
         public byte? Level { get; set; }
         public byte? Score { get; set; }
-        public byte? LeftTime { get; set; }
+        public int? LeftTime { get; set; }
     }
 }
